requirements.txt
discord.py
requests
discord_components
discord.py[voice]